Create LICENSE
Added licence with my name and 'et al.' - The original developers of the project aren't named on the Google Code archive, and there seems to be no claim that I can find to the project, so to still give credit to the original developers, et al. is used. If someone does know who these developers are, please get in touch so I can state this. Thanks!